farmconnect/server/ml/price_optimizer.py
import pandas as pd
from sklearn.ensemble import RandomForestRegressor
from joblib import dump, load
import os

class PriceOptimizer:
    def __init__(self):
        self.model_path = os.path.join(os.path.dirname(__file__), 'models/price_model.joblib')
        try:
            self.model = load(self.model_path)
        except:
            self.model = RandomForestRegressor(n_estimators=100)
    
    def train(self, X, y):
        """Train the price optimization model"""
        self.model.fit(X, y)
        dump(self.model, self.model_path)
        
    def predict_optimal_price(self, product_features):
        """Predict optimal price for a product"""
        df = pd.DataFrame([product_features])
        return self.model.predict(df)[0]
    
    def evaluate_market_trends(self, market_data):
        """Analyze market trends for pricing suggestions"""
        # Implementation would analyze competitor pricing, demand, etc.
        pass