farmconnect/server/requirements.txt
Django==4.2
djangorestframework==3.14.0
django-cors-headers==4.2.0
psycopg2-binary==2.9.6
Pillow==10.0.1
pandas==2.0.3
scikit-learn==1.3.0
joblib==1.3.2
python-dotenv==1.0.0
celery==5.3.1
redis==4.6.0